feat: #44 ExportImportsTemplate moved to Exports
