+++ src/Exports/ExportImportTemplate.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace HeadlessLaravel\Formations\Imports;
+namespace HeadlessLaravel\Formations\Exports;
 
 use HeadlessLaravel\Formations\Field;
 use HeadlessLaravel\Formations\Formation;
