+++ tests/ImportTest.php
@@ -2,7 +2,7 @@
 
 namespace HeadlessLaravel\Formations\Tests;
 
-use HeadlessLaravel\Formations\Imports\ExportImportTemplate;
+use HeadlessLaravel\Formations\Exports\ExportImportTemplate;
 use HeadlessLaravel\Formations\Mail\ImportErrorsMail;
 use HeadlessLaravel\Formations\Tests\Fixtures\Models\Post;
 use HeadlessLaravel\Formations\Tests\Fixtures\Models\User;
